requirements.txt
python-telegram-bot==20.7
anthropic==0.25.0
supabase==2.3.4
python-dotenv==1.0.0
httpx==0.27.0
